model-serving/api-gateway/requirements.txt
# API Gateway Dependencies
# FastAPI and ASGI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# Pydantic for data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP clients
httpx>=0.25.0
aiohttp>=3.9.0

# Kubernetes client
kubernetes>=28.1.0

# MLflow client
mlflow>=2.9.0

# Kubeflow Pipelines client
kfp>=2.4.0

# Redis client
redis>=5.0.0

# Authentication
PyJWT>=2.8.0
cryptography>=41.0.0

# Logging
structlog>=23.2.0
python-json-logger>=2.0.0

# Prometheus metrics
prometheus-client>=0.19.0

# Azure SDK (optional, for Azure AD)
azure-identity>=1.15.0
azure-storage-blob>=12.19.0

# Feast client (for feature retrieval)
feast>=0.34.0

# Data handling
pandas>=2.1.0
numpy>=1.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0
